feat: Finalized, i believe, DC Logger with modular architecture and better documentation
This commit adds the new structured logging system for Domo applications, featuring:
- Modular architecture with separate files for core components, configurations, and handlers. (all is separated, in smallcase for the files, PascalCase for the classes)
- Added some Factory functions for ease of configuration setup. (created a factory template)
- Enhanced the documentation of the readme and included an architecture.md for better understanding of the usage, we can talk over there.

Tried to follow SOLID Principles for this as best as i could.

One of the things to note on the changes was that all config classes method 'get_cloud_config()' was replaced with 'to_platform_config()' because cloud_config was misleading since not all configurations are cloud-based (console and file). This new name platform_config i feel it was better to describe the method return variable: configuration for the target platform/destination/handler.

The files changede are config/cloud.py and handlers/cloud/base.py

For your question Jae, the "Should ConsoleLogConfig inherit from LogConfig even though it's not a cloud config?"

My answer would be, Yes. All configurations share the same interface for polymorphism and the method rename i think should work better for this idea, so i changed the logic of the classes to match your suggestion i believe.

Sorry for the delay on the commit.